+++ src/Endpoint/ServicesInterface.php
@@ -7,6 +7,7 @@ namespace Aceproxies\ResellerApi\Endpoint;
 use Aceproxies\ResellerApi\Exception\ApiException;
 use Aceproxies\ResellerApi\Exception\InvalidResponseException;
 use Aceproxies\ResellerApi\Exception\TransportException;
+use Aceproxies\ResellerApi\Request\UpdateServiceRequest;
 use Aceproxies\ResellerApi\Response\ServiceDetailResponse;
 use Aceproxies\ResellerApi\Response\ServiceListResponse;
 use InvalidArgumentException;
@@ -28,4 +29,12 @@ interface ServicesInterface
      * @throws InvalidArgumentException
      */
     public function find(string $code): ?ServiceDetailResponse;
+
+    /**
+     * @throws ApiException
+     * @throws InvalidResponseException
+     * @throws TransportException
+     * @throws InvalidArgumentException
+     */
+    public function update(string $code, UpdateServiceRequest $request): void;
 }