+++ src/Response/EmptyResponse.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Aceproxies\ResellerApi\Response;
+
+final readonly class EmptyResponse
+{
+}